refactor: move Floor, Property, and Apartment types to /types and update imports in db.ts and SearchBar.jsx
