+++ src/data/db.ts
@@ -1,31 +1,4 @@
-export type Property = {
-  id: number
-  name: string
-  floors: Floor[]
-}
-
-export type Floor = {
-  id: number
-  name: string
-  apartments: Apartment[]
-}
-
-export type Apartment = {
-  id: number
-  status: string
-  propertyId: string
-  floor: string
-  address: string
-  city: string
-  rooms: number
-  bathrooms: number
-  size: number
-  rent: number
-  images: { id: number; url: string }[]
-  tenants: { id: number; name: string; email: string; leaseStart: Date }[]
-  description: string
-  issues: any[]
-}
+import { Property, Apartment } from '@/types/property'
 
 const properties: Property[] = [
   {
